+++ html/static/js/user/emptyengineer.js
@@ -2,6 +2,9 @@ $(document).ready(function() {
 
 var link='http://sw.womoe.top';
 $("#empty").click(function(event) {
+  if(!confirm("是否清空所有工程师")){
+    return ;
+  }
     $.ajax({
       url: link+'/hr/deleteAlll',
       type: 'GET',
